Manage multiple OpenNMS servers via profiles.
The initial version allowed users to save one server on a configuration
file. Now, it is possible to define multiple ones, and all the ReST
communication will happen against the default one.

When the tool starts, it will migrate the content if it finds the old
format on it, so that subsequent calls will have the new one.